+++ scripts/master.sh
@@ -26,8 +26,11 @@ mkdir -p /var/spool/torque/checkpoint/
 systemctl enable trqauthd
 systemctl start trqauthd
 
+
 # Configure pbs_server
+echo "Configure pbs_server"
 yes | pbs_server -t create
+sleep 5
 qmgr -c "set server acl_hosts=master"
 qmgr -c "set server scheduling=true"
 qmgr -c "create queue batch queue_type=execution"
@@ -39,12 +42,16 @@ qmgr -c "set server default_queue=batch"
 qmgr -c "set server keep_completed = 10"
 qterm
 
+
 # Build up the nodes file
+echo "Creating nodes file"
 : > /var/spool/torque/server_priv/nodes
 for i in `seq 1 $1`; do echo  "slave$i np=5" >> /var/spool/torque/server_priv/nodes; done
 
 # Set server_name to HOSTNAME
 echo $HOSTNAME > /var/spool/torque/server_name
+
+echo "Starting pbs_server"
 
 # Enable and start services
 systemctl enable pbs_server